Add missing DHCPv4 configuration and merge check mode function
DHCPv4 UseNTPServers configuration was missing in the configuration
function, but it was present in the check mode function. A dedicated
function only for check mode doesn't make sense because the logic is
basically the same.

Now there's only one configuration function and the missing DHCPv4
configuration is available if the module is run in check mode or not.