Add use_existing option to ipv4_address parameter
If you specify the wrong IPv4 settings, then the iLO won't be reachable
anymore. If you have a big environment with several iLO networks in the
same site, it can get tricky to assign the right iLO network to each iLO
interface in the inventory. You need to make sure to always assign the
right subnet mask and gateway. The IP address can be generated using the
inventory name.

Usually, if the IP address, subnet mask, and gateway are set to fixed,
it's not necessary anymore to set a different setting because the iLO
interface is reachable with these settings. With the use_existing option,
you skip the IPv4 address, subnet mask, and gateway configuration.